add magic

+++ noteql/__init__.py
@@ -8,6 +8,13 @@ import lxml.etree
 import html
 import pandas
 import xmltodict
+import pyparsing as pp
+import subprocess
+from collections import defaultdict
+from IPython.core.magic import (Magics, magics_class, line_cell_magic)
+from IPython.display import (display, HTML)
+from IPython import get_ipython
+
 
 LOCAL_DB_MADE = False
 
@@ -58,7 +65,6 @@ def create_local_db():
     global LOCAL_DB_MADE
     if LOCAL_DB_MADE:
         return
-    import subprocess
 
     cmd = """
         pip install --upgrade psycopg2-binary > pip.log
@@ -81,25 +87,30 @@ def create_local_db():
 
 def local_db_session():
     create_local_db()
-    return Session("public", "postgresql+psycopg2://root@/postgres")
+    return Session("postgresql+psycopg2://root@/postgres", "public")
 
 
 class Session:
-    def __init__(self, schema, dburi=None, drop_schema=False, overwrite=False):
+    def __init__(self, dburi, schema=None, drop_schema=False, overwrite=False):
         self.schema = schema
         self.dburi = dburi
         self.engine = get_engine(dburi)
         self.overwrite = overwrite
-        with self.engine.begin() as connection:
-            if drop_schema:
-                connection.execute(
-                    "drop schema if exists {} cascade;".format(self.schema)
-                )
-            connection.execute("create schema if not exists {};".format(self.schema))
+        if schema:
+            with self.engine.begin() as connection:
+                if drop_schema:
+                    connection.execute(
+                        "drop schema if exists {} cascade;".format(self.schema)
+                    )
+                connection.execute("create schema if not exists {};".format(self.schema))
+
+        ipython = get_ipython()
+        ipython.register_magics(Noteql)
 
     def get_results(self, sql, limit=-1):
         with self.engine.begin() as connection:
-            connection.execute("set local search_path = {};".format(self.schema))
+            if self.schema:
+                connection.execute("set local search_path = {};".format(self.schema))
             sql_result = connection.execute(sql)
             if sql_result.returns_rows:
                 results = {
@@ -116,8 +127,11 @@ class Session:
             return results
 
     def create_table(self, table, sql):
-        sql = '''
+        self.run_sql('''
             DROP TABLE IF EXISTS {table};
+        '''.format(table=table))
+
+        sql = '''
             CREATE TABLE {table}
             AS
             {sql}
@@ -134,26 +148,26 @@ class Session:
         chunksize=None,
     ):
         with self.engine.begin() as connection:
-            connection.execute("set local search_path = {};".format(self.schema))
+            if self.schema:
+                connection.execute("set local search_path = {};".format(self.schema))
             return pandas.read_sql_query(
                 sql,
                 connection,
                 index_col=None,
                 coerce_float=True,
-                params=None,
-                parse_dates=None,
-                chunksize=None,
+                params=params,
+                parse_dates=parse_dates,
+                chunksize=chunksize,
             )
 
     def show_dataframe(self, sql, title=None, title_size='h3', **kwargs):
-        from IPython.display import display
-        from IPython.display import HTML
-
         if title:
-            display(HTML(f'</br></br><{title_size}>{title}<{title_size}/>'))
+            self.show_title(title)
 
         display(self.get_dataframe(sql, **kwargs))
-        
+
+    def show_title(self, title):
+        display(HTML(f'</br></br><{title_size}>{title}<{title_size}/>'))
 
     def load_json(
         self,
@@ -189,7 +203,8 @@ class Session:
             full_name = schema_name + "." + table_name
 
         with self.engine.begin() as connection:
-            connection.execute("set local search_path = {};".format(self.schema))
+            if self.schema:
+                connection.execute("set local search_path = {};".format(self.schema))
             ## remove quotes when looking at actual table.
             if (
                 table_exists(connection, table_name[1:-1], schema_name[1:-1])
@@ -261,7 +276,8 @@ class Session:
             full_name = schema_name + "." + table_name
 
         with self.engine.begin() as connection:
-            connection.execute("set local search_path = {};".format(self.schema))
+            if self.schema:
+                connection.execute("set local search_path = {};".format(self.schema))
             if (
                 table_exists(connection, table_name[1:-1], schema_name[1:-1])
                 and not overwrite
@@ -327,7 +343,8 @@ class Session:
         method=None,
     ):
         with self.engine.begin() as connection:
-            connection.execute("set local search_path = {};".format(self.schema))
+            if self.schema:
+                connection.execute("set local search_path = {};".format(self.schema))
             dataframe.to_sql(
                 table_name,
                 connection,
@@ -417,3 +434,131 @@ class DecimalEncoder(json.JSONEncoder):
         if isinstance(obj, decimal.Decimal):
             return float(obj)
         return json.JSONEncoder.default(self, obj)
+
+
+@magics_class
+class Noteql(Magics):
+
+    def get_parsers(self):
+        arg = (pp.Word(pp.alphanums) + pp.Suppress("=") + (pp.Word(pp.alphanums) | pp.QuotedString("'", escQuote="'", unquoteResults=False)))("arg_params")
+        show = pp.Keyword("show", caseless=True)("show")
+        create = (pp.Keyword("create", caseless=True).suppress()  + (pp.Word(pp.alphanums+"_") | pp.QuotedString('"', escQuote='"', unquoteResults=False))("create"))
+        df = (pp.Keyword("df", caseless=True).suppress() + pp.Word(pp.alphanums+"_")("df"))
+        params = (pp.Keyword("params", caseless=True).suppress() + pp.Word(pp.alphanums+"_")("params"))
+        rest = pp.Word(pp.printables)("rest") 
+
+        magic_line_parser = pp.ZeroOrMore(pp.Group(arg | create | df | params | show | rest), stopOn=pp.LineEnd())
+
+        cell_parser = pp.OneOrMore(pp.SkipTo((pp.LineStart() + pp.Group(pp.Keyword("%%nql") + magic_line_parser)) | pp.StringEnd(), include=True))
+
+        return magic_line_parser, cell_parser
+
+    def execute_part(self, parsed_line, sql):
+        ns = self.shell.user_ns
+ 
+    
+        sql = sql.format(**ns)
+        
+        for key, value in ns.items():
+            if isinstance(value, Session):
+                session = value
+                break
+        else:
+            print('Need to define noteql seesion')
+               
+        actions = {}
+        params = {}
+        
+        for item in parsed_line:
+         
+            if item.getName() == 'rest':
+                print(f'Syntax error in %%nql not expecting {item}')
+                return
+            
+            if item.getName() == 'params':
+                param_name = item[0]
+                variable = ns.get(param_name)
+                if not variable:
+                    print(f'Error in %%nql, variable {param_name} does not exist in your notebook')
+                    return
+                params = variable
+
+            if item.getName() == 'arg_params':
+                param_name, param_value = item
+                if param_value[0] == "'":
+                    params[param_name] = param_value[1:-1]
+                else:
+                    variable = ns.get(param_value)
+                    if not variable:
+                        print(f'Error in %%nql, variable {param_name} does not exist in your notebook')
+                        return
+                    if params:
+                        print('Error in %%nql, can not use supplied params with the `params` argument')
+                        return
+                    params[param_name] = variable
+                    
+            if item.getName() == 'df':
+                if 'df' in actions:
+                    print(f'Error in %%nql, multiple df statements specified')
+                    return
+                actions['df'] = item[0]            
+            if item.getName() == 'create':
+                if 'create' in actions:
+                    print(f'Error in %%nql, multiple create statements specified')
+                    return
+                actions['create'] = item[0]
+
+            if item.getName() == 'show':
+                actions['show'] = True
+        
+     
+        if not actions:
+            actions['show'] = True
+
+        df = None
+        
+        df_name = actions.get('df')
+        show = 'show' in actions 
+        
+        if show or df_name:
+            if not params:
+                params=None
+            df = session.get_dataframe(sql, params=params)
+            if show:
+                display(df)
+            if df_name:
+                ns[df_name] = df
+        
+        create_name = actions.get('create')
+        
+        if create_name:
+            session.create_table(create_name, sql)
+                 
+    @line_cell_magic
+    def nql(self, line, cell=None):
+        
+        if cell:
+            magic_line_parser, cell_parser = self.get_parsers()
+
+            parsed_line = magic_line_parser.parseString(line)
+            parsed_cell = cell_parser.parseString(cell)
+
+            self.execute_part(parsed_line, parsed_cell[0])
+
+            for parsed_line, sql in zip(parsed_cell[1::2], parsed_cell[2::2]):
+                if not sql.strip():
+                    print(f'Error in %%nql, empty sql statement')
+                self.execute_part(parsed_line[1:], sql)
+        else:
+            ns = self.shell.user_ns
+            sql = line.format(**ns)
+    
+            for key, value in ns.items():
+                if isinstance(value, Session):
+                    session = value
+                    break
+            else:
+                print('Need to define noteql seesion')
+                return
+            return session.get_dataframe(sql)            
+
